fix: use Perl for safe NATS config variable substitution
The previous sed-based approach failed because sed has issues with
special characters in the bcrypt hash ($ and /).

Switch to Perl which handles replacement strings safely:
- Perl's s/// operator doesn't interpret special chars in replacement
- Properly handles $ and / in the bcrypt hash without escaping issues
- More reliable than sed for this use case

This fixes the "variable reference for '\$2y\$05\$...' can not be found" error.

Perl is available in all standard Unix/Linux environments used in CI/CD.